Updated README template to reflect Code.gov
Updated the README Template to reflect Code.gov requirements.

For the License File example - I think we need to update that file with Apache 2.0. From my understanding, that is the most popular open license on ITS CodeHub. If it is something else, then perhaps we should use that file. However, I do not think the MIT License is the best one to showcase for this file.

We also need to review the License and Contributing files to make sure they are compliant with the Code.gov requirements I added.

I deleted the "Code.gov Registration Info," as I tried to work these requirements into the entire template, with the idea that we can pull out the relevant sections for our Code.gov reporting and not ask users for information twice. However, I noticed there are some elements in the "Code.gov Registration Info" section that I did not notice as a Code.gov requirement. Namely:
- Tags: - my understanding is we go through a separate metadata process for this, so not sure we need to capture in the README (verified with Mike)
- Labor Hours: - not needed (verified with Mike)

If there is a reason to keep the above items, let me know.

Finally, because GitHub does not have spellcheck, in the future I think it should be part of the standard operating procedure to copy-paste this type of document into Microsoft Word, run spellcheck, and fix any typos in the README/template.